+++ RELEASENOTES.md
@@ -6,6 +6,7 @@
 *   ExoPlayer:
     *   Add support for using the virtual device ID from the `Context` passed to
         `ExoPlayer.Builder`.
+    *   Enable dynamic scheduling by default in scrubbing mode.
 *   Transformer:
     *   Add `CodecDbLite` that enables chipset specific optimizations of video
         encoding settings.